Week 2 Update
- Budget/fund generation.

- Monster, Earthquake, Tornado cannot damage anything within shield coverage.

- Shield overlay graph.

- Change in values for shield price, and fund.

- Change in city messages

+++ src/micropolisj/engine/BudgetNumbers.java
@@ -34,5 +34,5 @@ public class BudgetNumbers
 	
 	public int new_buildingRequest;
 	public int new_buildingFunded;
-	public int new_buildingPercent;
+	public double new_buildingPercent;
 }
